requirements.txt
flask
itsdangerous
flask-login
flask-sqlalchemy
passlib